+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.1] - 2016-03-04
+### Added
+- Added `AssertionFailedException` to be used with the `Assert\Assertion` class.
+- Adapted unit tests.
+
 ## [0.2.0] - 2016-02-18
 ### Fixed
 - Removed constructor from `ModuleExceptionTrait`. It was causing conflicts.
@@ -16,6 +21,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release to GitHub.
 
+[0.2.1]: https://github.com/brightnucleus/exceptions/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/brightnucleus/exceptions/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/brightnucleus/exceptions/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/brightnucleus/exceptions/compare/v0.0.0...v0.1.0